EDM-5220: Move delta worker handlers into tasks package
Restructure generate-delta logic to match other workers: consumer routing
lives in tasks/consumer.go and the sequential handler in
tasks/generate_delta.go, with matching consumer_test and
generate_delta_test files.

Made-with: Cursor
