+++ test/integration/delta_worker/consumer_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/flightctl/flightctl/internal/config"
 	"github.com/flightctl/flightctl/internal/consts"
-	deltaworker "github.com/flightctl/flightctl/internal/delta_worker"
+	deltatasks "github.com/flightctl/flightctl/internal/delta_worker/tasks"
 	"github.com/flightctl/flightctl/internal/domain"
 	"github.com/flightctl/flightctl/internal/worker_client"
 	flightlog "github.com/flightctl/flightctl/pkg/log"
@@ -86,7 +86,7 @@ var _ = Describe("Delta worker consumers", func() {
 		It("should consume and ack PrepareDeltas on DeltaGenerationTaskQueue", func() {
 			cfg := config.NewDefault()
 			cfg.DeltaGeneration = &config.DeltaGenerationConfig{MaxConcurrentDeltaGenerations: 1}
-			Expect(deltaworker.LaunchConsumers(ctx, provider, cfg, nil, nil, log)).To(Succeed())
+			Expect(deltatasks.LaunchConsumers(ctx, provider, cfg, nil, nil, log)).To(Succeed())
 
 			payload := prepareDeltasPayload()
 			producer, err := provider.NewQueueProducer(ctx, consts.DeltaGenerationTaskQueue)
@@ -102,7 +102,7 @@ var _ = Describe("Delta worker consumers", func() {
 		It("should not consume the same payload from TaskQueue", func() {
 			cfg := config.NewDefault()
 			cfg.DeltaGeneration = &config.DeltaGenerationConfig{MaxConcurrentDeltaGenerations: 1}
-			Expect(deltaworker.LaunchConsumers(ctx, provider, cfg, nil, nil, log)).To(Succeed())
+			Expect(deltatasks.LaunchConsumers(ctx, provider, cfg, nil, nil, log)).To(Succeed())
 
 			payload := prepareDeltasPayload()
 			producer, err := provider.NewQueueProducer(ctx, consts.TaskQueue)
